agent/requirements.txt
## NOTE: THIS REQUIREMENTS FILE IS JUST FOR LANGGRAPH STUDIO
## IT IS NOT INTENDED TO BE USED FOR LOCAL DEVELOPMENT

aiohttp~=3.10.0
arxiv~=2.1.3
duckduckgo-search~=6.2.6
langchain-community~=0.2.11
langchain-core~=0.2.26
langchain-openai~=0.1.20
langchain-groq~=0.1.9
numexpr~=2.10.1
python-dotenv~=1.0.1
# added by xroger88
tavily-python~=0.3.3
langchain-chroma~=0.1.3
chromadb~=0.5.3
pymongo~=4.8.0
